requirements.txt
Django
djangorestframework==3.9.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
defusedxml==0.6.0
django-allauth==0.39.1
django-cors-headers==2.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
JPype1==0.7.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.7
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
django_filter==2.2.0
django-cleanup==4.0.0
chatterbot
textblob
chatterbot-corpus
django chatterbot
